Licence number will show direct debit sales
https://eaflood.atlassian.net/browse/IWTF-1815

Currently, when doing a POCL import, the channel character in the
licence (e.g. 00011021-2WC3CMW-D28CA5, the second character 'W' in
the middle section represents a Web sale) is automatically set to
'P'.

This change will still default to that, but add a new character 'D'
if this is reflected in the imported file