Reinstate home countries in address entry
https://eaflood.atlassian.net/browse/IWTF-2069

Paragon only send bilingual licences to licensees with the country field
set to 'Wales'. This is set automatically when the postcode lookup is
used, however many Welsh postcodes result in an error from the lookup
service (see IWTF-2072). Therefore, customers living in Wales need to be
able to select 'Wales' in the manual address entry and are forced to
choose 'United Kingdom', meaning they can never receive a bilingual
licence