Update midnight start times
https://eaflood.atlassian.net/browse/IWTF-2465

We get comment from confused users about licence start times when the licence starts at the beginning of the day. Currently we say that licences start: '12:00am (midnight)'.

This task is to update the content of this specific start time so that it says: '0.00am (first minute of the day)' so that the time is clearer for users.