Generic Notice of Service Error
https://eaflood.atlassian.net/browse/IWTF-2524

This work is to improve the error screen so that it's possible to go
back to the step you were on previously. Also, to enable testing, there
needs to be a way of triggering a service error via setting
a querystring value which should be hidden behind a feature toggle so
that it doesn't make it to production.