Do not cache licensee (#989)
Do not cache licensee

https://eaflood.atlassian.net/browse/IWTF-2564

Second attempt at this. The previous solution to this didn't work as it was based in the updateTransaction routine, that runs after the POST handler so can't determine what the licence-for value was set to before the cache was updated to match the POSTed payload. I've moved this back into a custom POST handler.