Amend error page after going back during payment (#1306)
Amend error page after going back during payment

https://eaflood.atlassian.net/browse/IWTF-2867

A change to provide a clearer error page when an angler presses back to return from the payment pages to the Web Sales journey so they can be better informed about the courses of action open to them. To this end, I've amended the client error page to present more options when pressing back from the payment pages.