Fulfilment socket hangup remedy (#1318)
* Fulfilment socket hangup remedy

https://eaflood.atlassian.net/browse/IWTF-2836

A fix for the semi-regular socket hangups the fulfilment job experiences, causing records to be missed from the processing but be
marked as processed. This was caused by an incorrect calculation of part file number when the part file before a socket error hadn't been filled (number of fulfilment requests was lower than the max part file size), so we now detect that and correct the part file number.