Use chosen language when moving from renewals to new licence journey
https://eaflood.atlassian.net/browse/IWTF-2899

The "get a new fishing licence" link from the renewals journey was not persisting the user's chosen language - for example, if the user had previously selected Welsh, the link swapped them back to English. This PR fixes that.

It also includes some test refactoring to break up some of the tests to be a bit more granular.