Pin Docker to localstack v1.2.0
Localstack v1.3.0 was released yesterday. This included some breaking changes for us - the `DEFAULT_REGION` env var has been deprecated and the location of the init scripts appears to have changed. I wasn't able to immediately fix this so pinning to an earlier working version for now, as I think we'll need a Jira ticket to properly upgrade it.