Refactor redirects to automatically use correct language
https://eaflood.atlassian.net/browse/IWTF-3111

This PR refactors internal redirects to use a new redirectWithLanguageCode function on the response toolkit. This function allows us to call a redirect for the correct language without having to also call addLanguageCodeToUri each time.  This approach was initially demoed in https://github.com/DEFRA/rod-licensing/pull/1514 but this PR handles the implementation of the function and rolling it out to the rest of the app.