Strip existing lang params from the provided path
The previous attempt to fix this only worked if the argument was `lang=cy`. But if the path had `lang=en` we still want to be able to replace that with `lang=cy`. So this implementation just strips the existing language argument from the path so we can add a new one as needed.