Stop language parameters from being duplicated in URLs (#1603)
https://eaflood.atlassian.net/browse/IWTF-3111

Testing revealed that the links to accept and reject cookies were failing when the language was set to Welsh. This turned out to be because multiple `lang` parameters were being added - for example, `/buy?lang=cy&lang=cy`

We don't want this to happen, so this PR changes the addLanguageCodeToUri function to strip any existing `lang=` params from the provided URL so we can make sure we only add the relevant ones.