Amend permission end date to use locale time (#1651)
* Amend permission end date to use locale time

https://eaflood.atlassian.net/browse/IWTF-3324

The end date calculation for a permission is going wrong after the
change to BST, where it is going back an extra day. This is because the
permission start date is set as midnight in the locale time, which is
11pm the day before in GMT, and the end date calculation takes the last
minute of the previous date. Therefore '2023-03-31T23:00:00.000Z', which
is '2023-04-01T00:00:00.000+01:00' is resulting in an end date of
'2023-03-30T23:59:00.000Z'. It should actually be
'2023-03-31T22:59:00.000Z', or '2023-03-31T23:59:00.000+01:00'.