Use new isPhysical function on licence type page
To do this, we have to call findPermit first, but fortunately this is the first page in the journey where we should have all the necessary info to determine the permit.

This commit also adds a lot of tests for update-transaction.js as it was a bit of a black box before.

This change had a knock-on effect on a lot of seemingly-unrelated integration tests, which were hitting this page before testing their own pages and thus were now calling the sales API.