Display catch return info if any salmon licences are purchased
https://eaflood.atlassian.net/browse/IWTF-3316

This fixes a bug in multibuy where we were only displaying the catch return info on the order complete page if the most recently added licence was for salmon fishing.

This info should be displayed if any of the licences added are for salmon fishing, not just the latest one.