Old prices being sent to Govpay
https://eaflood.atlassian.net/browse/IWTF-3509

Old prices are being sent to Govpay because the start date isn't defined
when transactions are created for 'immediate start' permissions
(actually starting in 30 mintutes), because it's not possible to
determine how long someone might take to pay (or might decide not to buy
a licence after all). Setting a permission start time when the
transaction was created would also open a loophole where an angler could
start the purchase, then only compete it if they saw a bailiff coming,
receiving as a start time for their license whatever time they begun the
transaction.