Only prefill payment details when user buys for themself
https://eaflood.atlassian.net/browse/IWTF-3543

If a user is buying on behalf of someone else, we do not want to pre-populate GOV.UK Pay fields with the licensee info, as the person who is buying on behalf may be paying instead.

This PR changes the code to only pre-populate those fields when a user is buying a single licence for themself.

It also includes some test refactoring for the payment processor to bring it more in line with the rest of the test suite.