Stop easy renewals from copying shortTermPreferredMethodOfConfirmation
https://eaflood.atlassian.net/browse/IWTF-3614

This is causing a bug which causes easy renewals for contacts who have already got this field set in the CRM to fail, as the Sales API does not expect to be receiving the field in the data from the GAFL Web App.

By not copying the field when we do an easy renewal, we can leave the Sales API to assign the field itself as appropriate.