Fix issues with processing postal order records
https://eaflood.atlassian.net/browse/IWTF-3418

This PR fixes a few issues that came up while adding postal order records via the new CRM process.

There were a few mentions of fields prefixed with new* which had since been renamed - so these needed to be updated here as well.

Once this was changed, further errors revealed that we were passing through objects with labels rather than just the label string, which is what the sales API expected. So this PR also includes tweaks to make sure we pass through just a string when that is expected.