Strip fragments when processing URIs
https://eaflood.atlassian.net/browse/IWTF-3673

Selecting the 'skip to main content' link will append a #main-content fragment to the URL of the page. However this fragment then persists through the rest of the journey, which basically means that if you skipped once, you are forced to skip for the rest of the journey.

This PR updates how we process URIs to remove any fragments as well as adding or retaining the language code.