Stop redirect from always preserving skip link fragment (#1799)
https://eaflood.atlassian.net/browse/IWTF-3673

This change fixes an accessibility issue in which the #main-content fragment, which was applied when a user selected the 'Skip to main content' link, was then applied to all subsequent pages in the journey. This is because the 302 redirects always preserved the existing fragment.

As other possible solutions like updating the action tag then broke the language param, we decided to fix this by making sure a blank fragment is applied on redirects to override #main-content.