https://eaflood.atlassian.net/browse/IWTF-3690 (#1907)
Changed Cookie message button to be more accessible.

https://eaflood.atlassian.net/browse/IWTF-3690

A recent accessibility assessment identified a number of issues. This ticket is to rectify that labels were not descriptive of their topic or purpose in the cookies acceptance banner in the GAFL journey. The issue and suggested solution is detailed on page 66-67 of the assessment.