Rectify the functionality of the word "Cymraeg" in the language toggle.
https://eaflood.atlassian.net/browse/IWTF-3689

A recent accessibility assessment identified a number of issues. This ticket is to rectify the functionality the word “Cymraeg” in the language toggle in the GAFL journey.

The issue and suggested solution is detailed on page 62-65 of the assessment.

The human language of the element is correctly set using the lang attribute to its corresponding ISO language code on every page. The correct lang attribute is set on the language link for Cymraeg when the page is in English, and the English link when the page is in Welsh