Hiding and revealing content on licence type selection (#1916)
Hiding and revealing content on Licence type selection

https://eaflood.atlassian.net/browse/IWTF-3809

A recent accessibility assessment identified a number of issues. This ticket is to make the page for the screen reader more accessible when hiding and revealing content based on the licence type option selected. This issue and suggested solution is detailed on page 68-71 of the DAC assessment.