Improving usability to page title to meet WGAC 2.1 criteria (#1920)
Improving usability to page title to meet WGAC 2.1 criteria

https://eaflood.atlassian.net/browse/IWTF-3824

A recent accessibility assessment identified a number of issues. This ticket is to make the Page titles meet the criteria of the WCAG 2.1 criteria. The issue and suggested solution is detailed on page 91-93 of the DAC Assessment.

This issue affects all users, but in particular, screen reader users who are dependant on having unique and accurate page titles to orient themselves within the
 service.