Adding in “Error” prefix in titles
feature/iwtf-3830-error-prefix-in-title

This ticket is to make the Page title match the Error handling pattern of the GOV.UK Design.

The issue and suggested solution is detailed on page 99-101 of the DAC assessment.

On the ‘Who is this fishing licence for?’ page, when an error occurs, the page title does not include an ‘Error:’ prefix, that is indicated in the GOV.UK Design System error handling pattern.

 This affects screen reader users in particular, as the ‘Error:’ prefix helps them to understand that they’ve been returned back to the same page and why they have not changed page, upon submitting the form.