Keyboard focus inconsistency (#1925)
https://eaflood.atlassian.net/browse/IWTF-3838

A recent accessibility assessment identified a number of issues. This ticket is to make keyboard focus consistent on Change links with the GOV.UK Design System. The issue and suggested solution is detailed on page 118-119 of the DAC assessment.

This affects keyboard only users, as this link was styled inconsistently from other links, using the browser default styling. It now uses the yellow and black, thick underline, styling used everywhere else.