Remove "Continue" button from within the fieldset (#1938)
https://eaflood.atlassian.net/browse/IWTF-3833

This ticket is to reposition Continue button outside the field set. The issue and suggested solution is detailed on page 110-111 of the DAC assessment.

On the ‘Find your address’ page, the ‘Continue’ submit button for the form has been included inside the fieldset.

This affects screen reader users as screen readers announce the start and end of fieldsets, in order to assist the user in understanding the structure of the page. Including the ‘Continue’ button within the group could lead to the implication that the button is part of the question, not a means of submitting the form.