Associating hint text with the field set.
https://eaflood.atlassian.net/browse/IWTF-3831

A recent accessibility assessment identified a number of issues. This ticket is to make the hint text to associate with the field set. The issue and suggested solution is detailed on page 108-109 of the DAC assessment.

On the ‘Who is this fishing licence for?’ page, the hint text provided is not associated with the field set.

 This affects screen reader users as the visual relationship of the hint text, styled as hint text, is not equally reflected programmatically for them.