Ensuring page titles meet WCAG 2.1 criteria (#1940)
* Ensuring page titles meet WCAG 2.1 criteria

https://eaflood.atlassian.net/browse/IWTF-3824
A recent accessibility assessment identified a number of issues. This ticket is to make the Page titles meet the criteria of the WCAG 2.1 criteria. The issue and suggested solution is detailed on page 91-93 of the DAC assessment.
On the ‘Who is this fishing licence for?’ page, the page title does not follow the GOV.UK Design System page title recommendations; including "Get a rod fishing licence" helps screen reader users, who are dependent on having unique and accurate page titles to orientate themselves within the service.

* Ensuring payment cancelled page follows revision.

* Changes to check-confirmation completed.

* Revert check-confirmation-contact.njk

* Update contact.njk

* Update check-confirmation-contact.njk

* Correct address-lookup.njk merge error

* Correct contact.njk merge error

* Correct terms-and-conditions.njk merge error

* Correct identify.njk merge error

* Update identify.njk

* Update address-lookup.njk

* Correct check-confirmation-contact.njk merge error

* Correct choose-payment.njk merge error

* Correct terms-and-conditions.njk merge error

* Correct choose-payment.njk merge error

* Update check-confirmation-contact.njk

* Correct choose-payment.njk merge error