Amend refund policy page
https://eaflood.atlassian.net/browse/IWTF-3662

This ticket's purpose is to add an email to the refund policy page, to allow users to request a refund by email. There is no risk of refund being refused for the user because the email sent date is taken as the being within the 14 day refund period - not the ‘read by staff’ date.