Fixing privacy title discrepancy (#1945)
* https://eaflood.atlassian.net/browse/IWTF-4062

A bug was discovered on the privacy policy page. The Welsh page title matched the H1 title in English, but was different in Welsh. The page titles should match the H1 titles.

* Update cy.json

Updated to ensure I wasn't changing the Welsh title of the service across the whole service, and only for this page specifically.