Correcting details on Welsh payment screen
https://eaflood.atlassian.net/browse/IWTF-3981

In the Welsh user journey, GOV.UK PAY defaults correctly to Welsh, pulling through the details of purchase from the main service. However, in at least one instance, we are supplying the licence length in English – see attached image. In the grey box, it says "Brithyllod a physgod bras, hyd at 2 wialen, 1 day" instead of "1 diwrnod".

This ticket is to ensure that details are correctly supplied in Welsh for all licence types.