Recurring payments job sets up new permissions in DynamoDb
https://eaflood.atlassian.net/browse/IWTF-4017

When the RP job is run and relevant permissions are found that need to be renewed, we want to create a new permission for the renewal and store it in DynamoDb. This permission should have all the relevant data that needs to be carried over. The new permission won't be moved over to the CRM until payment has successfully been taken.

As part of this we need to adapt some of the code that was written for easy renewals.