Add new route to Sales API connector
https://eaflood.atlassian.net/browse/IWTF-4017

This is required for IWTF-4017 as we will need a new function within the Sales API to prepare data for a recurring payment. The recurring payment job needs to be able to conenct to this new feature so this PR updates the connector. See https://github.com/DEFRA/rod-licensing/pull/1957/ for further implementation.