Add agreement_id to existing GOV.UK Pay request
https://eaflood.atlassian.net/browse/IWTF-4261

If the request to prepare a payment is part of setting up a new recurring payment agreement, then the agreement_id should be included.

I also renamed prepareRecurringPayment() to prepareRecurringPaymentAgreement() as I think this is slightly more accurate - it is setting up the agreement rather than the payment - and it initially threw me off when working out what needed to be changed.