Clarify RP agreement status when first payment does not go through
https://eaflood.atlassian.net/browse/IWTF-4514

When a user decides to set up a recurring payment agreement but then cancels the first payment, or the first payment fails, we want to include additional text on the cancel or failure page to confirm that they have not finished setting up the agreement for future payments.

This includes a bit of refactoring to use some existing logic from the order complete page.