https://eaflood.atlassian.net/browse/IWTF-4392
The easy renewals journey is using error validation to tell a user they are not eligible to use the service (in this case because their details do not match any in our systems). This is not a correct use of the design pattern: Recover from validation errors

“Do not use validation to check whether the user is eligible to use the service”

Instead the user should be taken to a new screen that tells them what has happened and what they need to do next.

The main problems are:

not accessible for screen readers - error at top of page and above field do not match

because we cannot tell users which piece of information is incorrect for security reasons, the error is always applied to the licence number (from what I can tell from my limited testing). The user may waste time and get frustrated re-checking the licence number, not realising it is another field that is incorrect.