RP Job Handles API erros / failed payments
https://eaflood.atlassian.net/browse/IWTF-4021

When the recurring payments job sends a “create payment” request to GOV.UK Pay , we expect it to return confirmation that our payment request has been received.

Similarly, when we send our second request , we expect it to return the status of our payment request.

However, if our request to the payment API returns an error response code (for example, if the API is down or if our authorisation is invalid) then the recurring payment job needs to handle this gracefully.

AC

If we receive an error response, this response is outputted to our logs

Abort the run if systems are down at the start of a run and try again

Continue with run if systems are down mid-run and try again

“Trying again” means we will have a second run during the day (first run in the morning at a different time to the POCL run,  second run in the afternoon at 15:00). Any failed payments from the morning run will be picked up in the afternoon run.

If, after both runs, a payment has still not been taken,  the run on the following day will detect that the window for processing these payments has passed and automatically cancel them.