Add new endpoint to link recurring payments
We need to have the id of the new recurring payment before we can attach it to the old recurring payment, so I've decided to split this off into a separate Sales API call that will be made from the job after the new recurring payment records are created. This commit sets up an extremely basic endpoint in the Sales API to get the new connector in place, while the next PR will add functionality to the Sales API and then call it from the RP job.