Update disability concession eligibility
https://eaflood.atlassian.net/browse/IWTF-4672

Following challenge, the business are changing the Blue badge requirement for disabled concessionary licences.

A customer must now meet the eligibility requirements of a blue badge, but no longer needs to actually hold one. This is to remove the necessity for a customer to apply for one permit just to allow them to apply for a second (our licence).

A customer no longer has to supply  a Blue badge number to prove they have one. This removes the risk of collecting personal data that we do not use.