RP job handles GOV.UK Pay error responses
https://eaflood.atlassian.net/browse/IWTF-4021

When the recurring payments job sends a “create payment” request to GOV.UK Pay. , we expect it to return confirmation that our payment request has been received.
However, if our request to the payment API returns an error response code (for example, if the API is down or if our authorisation is invalid) then the recurring payment job needs to handle this gracefully.