RCP - Clean up from previous jobs (#2183)
* RP job handles GOV.UK Pay error responses

https://eaflood.atlassian.net/browse/IWTF-4021

When the recurring payments job sends a “create payment” request to GOV.UK Pay. , we expect it to return confirmation that our payment request has been received.
However, if our request to the payment API returns an error response code (for example, if the API is down or if our authorisation is invalid) then the recurring payment job needs to handle this gracefully.

* Amend RP job to create transaction with RP id _and_ agreement id
* Refactored payment result processing and added call to cancel recurring payments where necessary
* Refactor tests for payment mop-up processor, getting fresh data for tests when they use the standard journal entries, and mocking the Bottleneck dependency to speed up test execution
* Refactor payment mop up processor tests as there were several repeated tests not adding value, as well as some unnecessary setup
* Tidy up tests to avoid persistent mock implementations

---------

Co-authored-by: laila aleissa <138867360+lailien3@users.noreply.github.com>
Co-authored-by: Scott <scott.dormand@defra.gov.uk>