Recurring payment job should handle errors from await salesApi.processRPResult() (#2215)
https://eaflood.atlassian.net/browse/IWTF-4649

Currently, when the recurring payment job calls await salesApi.processRPResult, we await the response but we don’t do anything with it. We should have some error handling in place so that if one record fails to process, we can log it to be reviewed later, but also not have the whole job fall over.

* test branch due to rebase issues

* use mockResolvedValueOnce to avoid mocks spillign over

* tighten up and make pretty tests